app/requirements.txt
flask
flask_httpauth
mysql-connector-python
numpy
Pillow
tensorflow
torch
torchvision
python-dotenv